+++ R/dashboardNavbar.R
@@ -97,11 +97,13 @@ bs4DashNavbar <- function(..., skin = "light", status = "white", border = TRUE,
 #'
 #' Build an adminLTE3 dashboard dropdown menu
 #'
-#' @param ... Slot for bs4Notification or bs4Message.
+#' @param ... Slot for \link{bs4DropdownMenuItem}.
 #' @param show Whether to start with the dropdown open. FALSE by default.
 #' @param status Dropdown menu status. "primary", "success", "warning", "danger" or "info".
-#' @param labelText Dropdown label text.,
+#' @param labelText Dropdown label text.
 #' @param src Dropdown link to an external ressource.
+#' @param menuIcon Fontawesome icon (default = "bell")
+#' @param align Menu alignment (default = "right")
 #' 
 #' @examples
 #' if(interactive()){
@@ -140,7 +142,8 @@ bs4DashNavbar <- function(..., skin = "light", status = "white", border = TRUE,
 #'
 #' @export
 bs4DropdownMenu <- function(..., show = FALSE, labelText = NULL, src = NULL,
-                            status = c("primary", "warning", "danger", "info", "success")) {
+                            status = c("primary", "warning", "danger", "info", "success"), 
+                            menuIcon = "bell", align = "right") {
   
   status <- match.arg(status)
   items <- list(...)
@@ -148,13 +151,15 @@ bs4DropdownMenu <- function(..., show = FALSE, labelText = NULL, src = NULL,
   # remove the divider from the last item
   #items[[n_items]][[2]] <- NULL
   
+  labelText <- n_items
+  
   dropdownMenuTag <- shiny::tags$li(
     class = if (isTRUE(show)) "nav-item dropdown show" else "nav-item dropdown",
     shiny::tags$a(
       class = "nav-link",
       `data-toggle` = "dropdown",
       href = "#",
-      shiny::icon("bell"),
+      shiny::icon(menuIcon),
       shiny::tags$span(
         class = paste0("badge badge-", status, " navbar-badge"), 
         labelText
@@ -162,13 +167,13 @@ bs4DropdownMenu <- function(..., show = FALSE, labelText = NULL, src = NULL,
     ),
     shiny::tags$div(
       class = if (isTRUE(show)) {
-        "dropdown-menu dropdown-menu-lg dropdown-menu-right show"
+        sprintf("dropdown-menu dropdown-menu-lg dropdown-menu-%s show", align)
       } else {
-        "dropdown-menu dropdown-menu-lg dropdown-menu-right"
+        sprintf("dropdown-menu dropdown-menu-lg dropdown-menu-%s", align)
       },
       shiny::tags$span(
         class = "dropdown-item dropdown-header", 
-        paste0(n_items, " Notifications")
+        paste0(n_items, " Items")
       ),
       shiny::tags$div(class = "dropdown-divider"),
       ...,
@@ -186,12 +191,12 @@ bs4DropdownMenu <- function(..., show = FALSE, labelText = NULL, src = NULL,
       shiny::tags$head(
         shiny::tags$style(
           shiny::HTML(
-            paste0(
-              ".fa-bell {
+            sprintf(
+              ".fa-%s {
                 color: #000;
                }
               "
-            )
+            , menuIcon)
           )
         )
       )