+++ inst/bs4Dash-0.2.0/bs4Dash.js
@@ -1,4 +1,14 @@
 $(function () {
+  
+  // Make the dashboard widgets sortable Using jquery UI
+  $('.connectedSortable').sortable({
+    placeholder         : 'sort-highlight',
+    connectWith         : '.connectedSortable',
+    handle              : '.card-header, .nav-tabs',
+    forcePlaceholderSize: true,
+    zIndex              : 999999
+  });
+  $('.connectedSortable .card-header, .connectedSortable .nav-tabs-custom').css('cursor', 'move');
 
   //// function to get the index of an element in a list
   //function getIndex(elem) {
@@ -17,21 +27,20 @@ $(function () {
   // handle shinyapps.io: w need to extract the worker id and
   // paste it in the url so that the apps works correctly
   // get the shiny app.io workerId
+  // handles shinyapps.io
   var workerId = $('base').attr('href');
-  // ensure that this code does not run on shiny server/pro and locally
+  // ensure that this code does not locally
   if (typeof workerId != "undefined") {
-    // get the initial page url
-    var url = window.location.href;
-    // get the name of the first selected tab
-    $tablink = $('#mymenu .nav-item:eq(0) a').attr('href');
-    // replace the url by the url for shinyapp.io
-    window.location.replace(url + workerId + $tablink);
+    var pathname = window.location.pathname;
+    var newpath = pathname + workerId;
+    console.log(newpath);
+    window.history.replaceState( {} , 'newpath', newpath);
   }
   
   if ($selectedTab.length === 0) {
     // If no tab starts selected, use the first one, if present
     $('#mymenu .nav-item:eq(0) a').tab('show');
-    $('.tab-pane:eq(0)').addClass('active show');
+    $('.container-fluid.tab-pane:eq(0)').addClass('active show');
   } //else {
     //$("#mymenu .nav-item:eq(" + $indexPane + ") a").tab("show");
     //$(".tab-pane:eq(" + $indexPane + ")").addClass('active show');